api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Taraslysak <lysak.ipr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
fastapi = "^0.88.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
flake8 = "^6.0.0"
black = "^22.12.0"
redis = "^4.4.0"
pytest-mock-resources = {extras = ["redis"], version = "^2.6.5"}
httpx = "^0.23.1"
passlib = "^1.7.4"
uuid = "^1.30"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
